1.新增二手相关接口 2.新增赢大奖相关接口

+++ ehome-api/activity-api/src/main/java/com/busi/entity/PrizesEvent.java
@@ -29,7 +29,7 @@ public class PrizesEvent {
 
     private int issue;		//期号
 
-    private double price;//奖品价值
+    private int price;//奖品价值
 
     private String imgUrl;			//奖品图
