+++ ehome-api/activity-api/src/main/java/com/busi/entity/PrizesMemorial.java
@@ -21,7 +21,7 @@ public class PrizesMemorial {
 
     private int issue;		//期号
 
-    private double price;//奖品价值
+    private int price;//奖品价值
 
     private String imgUrl;			//奖品图
 }
