+++ ehome-api/activity-api/src/main/java/com/busi/entity/PrizesReceipt.java
@@ -28,7 +28,7 @@ public class PrizesReceipt {
 
     private int issue;		//期号
 
-    private double price;//奖品价值
+    private int price;//奖品价值
 
     private String imgUrl;			//奖品图
